--- tests/integration.rs
@@ -1,1055 +1,0 @@
-use std::{borrow::Cow, fs::File, io::Write};
-
-use assert_cmd::Command;
-use mdtablefix::{
-    THEMATIC_BREAK_LEN,
-    convert_html_tables,
-    format_breaks,
-    process_stream,
-    reflow_table,
-    renumber_lists,
-};
-use rstest::{fixture, rstest};
-use tempfile::tempdir;
-
-#[macro_use]
-mod common;
-
-#[fixture]
-/// Provides a sample Markdown table with broken rows for testing purposes.
-///
-/// The returned vector contains lines representing a table with inconsistent columns, useful for
-/// validating table reflow logic.
-fn broken_table() -> Vec<String> { return lines_vec!["| A | B |    |", "| 1 | 2 |  | 3 | 4 |"]; }
-
-#[fixture]
-/// Returns a vector of strings representing a malformed Markdown table with inconsistent columns.
-///
-/// The returned table has rows with differing numbers of columns, making it invalid for standard
-/// Markdown table parsing.
-fn malformed_table() -> Vec<String> { return lines_vec!["| A | |", "| 1 | 2 | 3 |"]; }
-
-#[fixture]
-fn header_table() -> Vec<String> {
-    lines_vec!["| A | B |    |", "| --- | --- |", "| 1 | 2 |  | 3 | 4 |"]
-}
-
-#[fixture]
-fn escaped_pipe_table() -> Vec<String> {
-    lines_vec!["| X | Y |    |", "| a \\| b | 1 |  | 2 | 3 |"]
-}
-
-#[fixture]
-fn indented_table() -> Vec<String> {
-    return lines_vec!["  | I | J |    |", "  | 1 | 2 |  | 3 | 4 |"];
-}
-
-#[fixture]
-fn html_table() -> Vec<String> {
-    lines_vec![
-        "<table>",
-        "<tr><th>A</th><th>B</th></tr>",
-        "<tr><td>1</td><td>2</td></tr>",
-        "</table>",
-    ]
-}
-
-#[fixture]
-fn html_table_with_attrs() -> Vec<String> {
-    lines_vec![
-        "<table class=\"x\">",
-        "<tr><th>A</th><th>B</th></tr>",
-        "<tr><td>1</td><td>2</td></tr>",
-        "</table>",
-    ]
-}
-
-#[fixture]
-fn html_table_with_colspan() -> Vec<String> {
-    lines_vec![
-        "<table>",
-        "<tr><th colspan=\"2\">A</th></tr>",
-        "<tr><td>1</td><td>2</td></tr>",
-        "</table>",
-    ]
-}
-
-#[fixture]
-fn html_table_no_header() -> Vec<String> {
-    lines_vec![
-        "<table>",
-        "<tr><td>A</td><td>B</td></tr>",
-        "<tr><td>1</td><td>2</td></tr>",
-        "</table>",
-    ]
-}
-
-#[fixture]
-fn html_table_empty_row() -> Vec<String> {
-    lines_vec![
-        "<table>",
-        "<tr></tr>",
-        "<tr><td>1</td><td>2</td></tr>",
-        "</table>",
-    ]
-}
-
-#[fixture]
-fn html_table_whitespace_header() -> Vec<String> {
-    lines_vec![
-        "<table>",
-        "<tr><td>  </td><td>  </td></tr>",
-        "<tr><td>1</td><td>2</td></tr>",
-        "</table>",
-    ]
-}
-
-#[fixture]
-fn html_table_inconsistent_first_row() -> Vec<String> {
-    lines_vec![
-        "<table>",
-        "<tr><td>A</td></tr>",
-        "<tr><td>1</td><td>2</td></tr>",
-        "</table>",
-    ]
-}
-
-#[fixture]
-fn html_table_empty() -> Vec<String> { return lines_vec!["<table></table>"]; }
-
-#[fixture]
-fn html_table_unclosed() -> Vec<String> { return lines_vec!["<table>", "<tr><td>1</td></tr>"]; }
-
-#[fixture]
-fn html_table_uppercase() -> Vec<String> {
-    lines_vec![
-        "<TABLE>",
-        "<tr><th>A</th><th>B</th></tr>",
-        "<tr><td>1</td><td>2</td></tr>",
-        "</TABLE>",
-    ]
-}
-
-#[fixture]
-fn html_table_mixed_case() -> Vec<String> {
-    lines_vec![
-        "<TaBlE>",
-        "<tr><th>A</th><th>B</th></tr>",
-        "<tr><td>1</td><td>2</td></tr>",
-        "</TaBlE>",
-    ]
-}
-
-#[fixture]
-fn multiple_tables() -> Vec<String> {
-    lines_vec!["| A | B |", "| 1 | 22 |", "", "| X | Y |", "| 3 | 4 |"]
-}
-
-#[rstest]
-/// Tests that `reflow_table` correctly restructures a broken Markdown table into a well-formed
-/// table.
-fn test_reflow_basic(broken_table: Vec<String>) {
-    let expected = lines_vec!["| A | B |", "| 1 | 2 |", "| 3 | 4 |"];
-    assert_eq!(reflow_table(&broken_table), expected);
-}
-
-#[rstest]
-/// Tests that `reflow_table` returns the original input unchanged when given a malformed Markdown
-/// table.
-///
-/// This ensures that the function does not attempt to modify tables with inconsistent columns or
-/// structure.
-fn test_reflow_malformed_returns_original(malformed_table: Vec<String>) {
-    assert_eq!(reflow_table(&malformed_table), malformed_table);
-}
-
-#[rstest]
-fn test_reflow_preserves_header(header_table: Vec<String>) {
-    let expected = lines_vec!["| A | B |", "| --- | --- |", "| 1 | 2 |", "| 3 | 4 |"];
-    assert_eq!(reflow_table(&header_table), expected);
-}
-
-#[rstest]
-fn test_reflow_handles_escaped_pipes(escaped_pipe_table: Vec<String>) {
-    // The fixture contains a header row followed by a row with an escaped
-    // pipe sequence (`a \| b`). After reflow the escaped pipe becomes a literal
-    // `|` inside the first data cell, so the table has three columns and the
-    // header row is padded to match.
-    let expected = lines_vec!["| X     | Y |", "| a | b | 1 |", "| 2     | 3 |"];
-    assert_eq!(reflow_table(&escaped_pipe_table), expected);
-}
-
-#[rstest]
-fn test_reflow_preserves_indentation(indented_table: Vec<String>) {
-    let expected = lines_vec!["  | I | J |", "  | 1 | 2 |", "  | 3 | 4 |"];
-    assert_eq!(reflow_table(&indented_table), expected);
-}
-
-#[rstest]
-fn test_process_stream_html_table(html_table: Vec<String>) {
-    let expected = lines_vec!["| A | B |", "| --- | --- |", "| 1 | 2 |"];
-    assert_eq!(process_stream(&html_table), expected);
-}
-
-#[rstest]
-fn test_process_stream_html_table_with_attrs(html_table_with_attrs: Vec<String>) {
-    let expected = lines_vec!["| A | B |", "| --- | --- |", "| 1 | 2 |"];
-    assert_eq!(process_stream(&html_table_with_attrs), expected);
-}
-
-#[rstest]
-fn test_process_stream_html_table_uppercase(html_table_uppercase: Vec<String>) {
-    let expected = lines_vec!["| A | B |", "| --- | --- |", "| 1 | 2 |"];
-    assert_eq!(process_stream(&html_table_uppercase), expected);
-}
-
-#[rstest]
-fn test_process_stream_html_table_mixed_case(html_table_mixed_case: Vec<String>) {
-    let expected = lines_vec!["| A | B |", "| --- | --- |", "| 1 | 2 |"];
-    assert_eq!(process_stream(&html_table_mixed_case), expected);
-}
-
-#[rstest]
-fn test_process_stream_multiple_tables(multiple_tables: Vec<String>) {
-    let expected = lines_vec!["| A | B  |", "| 1 | 22 |", "", "| X | Y |", "| 3 | 4 |",];
-    assert_eq!(process_stream(&multiple_tables), expected);
-}
-
-/// Tests that `process_stream` leaves lines inside code fences unchanged.
-///
-/// Verifies that both backtick (```) and tilde (~~~) fenced code blocks are ignored by the table
-/// processing logic, ensuring their contents are not altered.
-#[rstest]
-fn test_process_stream_ignores_code_fences() {
-    let lines = lines_vec!["```rust", "| not | a | table |", "```"];
-    assert_eq!(process_stream(&lines), lines);
-
-    // Test with tilde-based code fences
-    let tilde_lines = lines_vec!["~~~", "| not | a | table |", "~~~"];
-    assert_eq!(process_stream(&tilde_lines), tilde_lines);
-}
-
-#[rstest]
-fn test_process_stream_ignores_indented_fences() {
-    let lines = lines_vec!(
-        "   ```javascript",
-        "   socket.onmessage = function(event) {",
-        "       const message = JSON.parse(event.data);",
-        "       switch(message.type) {",
-        "           case \"serverNewMessage\":",
-        "               // Display message.payload.user and message.payload.text",
-        "               break;",
-        "           case \"serverUserJoined\":",
-        "               // Update user list with message.payload.user",
-        "               break;",
-        "           // Handle other message types...",
-        "       }",
-        "   };",
-        "",
-        "   ```",
-    );
-    assert_eq!(process_stream(&lines), lines);
-}
-
-#[rstest]
-/// Verifies that the CLI fails when the `--in-place` flag is used without specifying a file.
-///
-/// This test ensures that running `mdtablefix --in-place` without a file argument results in a
-/// command failure.
-fn test_cli_in_place_requires_file() {
-    Command::cargo_bin("mdtablefix")
-        .unwrap()
-        .arg("--in-place")
-        .assert()
-        .failure();
-}
-
-#[rstest]
-/// Tests that the CLI processes a file containing a broken Markdown table and outputs the corrected
-/// table to stdout.
-///
-/// This test creates a temporary file with a malformed table, runs the `mdtablefix` binary on it,
-/// and asserts that the output is the expected fixed table.
-fn test_cli_process_file(broken_table: Vec<String>) {
-    let dir = tempdir().unwrap();
-    let file_path = dir.path().join("sample.md");
-    let mut f = File::create(&file_path).unwrap();
-    for line in &broken_table {
-        writeln!(f, "{line}").unwrap();
-    }
-    f.flush().unwrap();
-    drop(f);
-    Command::cargo_bin("mdtablefix")
-        .unwrap()
-        .arg(&file_path)
-        .assert()
-        .success()
-        .stdout("| A | B |\n| 1 | 2 |\n| 3 | 4 |\n");
-}
-
-#[test]
-fn test_cli_wrap_option() {
-    let input = "This line is deliberately made much longer than eighty columns so that the \
-                 wrapping algorithm is forced to insert a soft line-break somewhere in the middle \
-                 of the paragraph when the --wrap flag is supplied.";
-    let output = Command::cargo_bin("mdtablefix")
-        .unwrap()
-        .arg("--wrap")
-        .write_stdin(format!("{input}\n"))
-        .output()
-        .unwrap();
-    assert!(output.status.success());
-    let text = String::from_utf8_lossy(&output.stdout);
-    assert!(
-        text.lines().count() > 1,
-        "expected wrapped output on multiple lines"
-    );
-    assert!(text.lines().all(|l| l.len() <= 80));
-}
-
-#[test]
-fn test_uniform_example_one() {
-    let input = lines_vec![
-        "| Logical type | PostgreSQL | SQLite notes |",
-        "|--------------|-------------------------|---------------------------------------------------------------------------------|",
-        "| strings | `TEXT` (or `VARCHAR`) | `TEXT` - SQLite ignores the length specifier anyway |",
-        "| booleans | `BOOLEAN DEFAULT FALSE` | declare as `BOOLEAN`; Diesel serialises to 0 / 1 so this is fine |",
-        "| integers | `INTEGER` / `BIGINT` | ditto |",
-        "| decimals | `NUMERIC` | stored as FLOAT in SQLite; Diesel `Numeric` round-trips, but beware precision |",
-        "| blobs / raw | `BYTEA` | `BLOB` |",
-    ];
-    let output = reflow_table(&input);
-    assert!(!output.is_empty());
-    let widths: Vec<usize> = output[0]
-        .trim_matches('|')
-        .split('|')
-        .map(str::len)
-        .collect();
-    for row in output {
-        let cols: Vec<&str> = row.trim_matches('|').split('|').collect();
-        for (i, col) in cols.iter().enumerate() {
-            assert_eq!(col.len(), widths[i]);
-        }
-    }
-}
-
-#[test]
-fn test_uniform_example_two() {
-    let input = lines_vec![
-        "| Option | How it works | When to choose it |",
-        "|--------------------------------------|--------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------|------------------------------------------------------------------------------------------------|",
-        "| **B. Pure-Rust migrations** | Implement `diesel::migration::Migration<DB>` in a Rust file (`up.rs` / `down.rs`) and compile with both `features = [\"postgres\", \"sqlite\"]`. The query builder emits backend-specific SQL at runtime. | You prefer the type-checked DSL and can live with slightly slower compile times. |",
-        "| **C. Lowest-common-denominator SQL** | Write one `up.sql`/`down.sql` that *already* works on both engines. This demands avoiding SERIAL/IDENTITY, JSONB, `TIMESTAMPTZ`, etc. | Simple schemas, embedded use-case only, you are happy to supply integer primary keys manually. |",
-        "| **D. Two separate migration trees** | Maintain `migrations/sqlite` and `migrations/postgres` directories with identical version numbers. Use `embed_migrations!(\"migrations/<backend>\")` to compile the right set. | You ship a single binary with migrations baked in. |",
-    ];
-    let output = reflow_table(&input);
-    assert!(!output.is_empty());
-    let widths: Vec<usize> = output[0]
-        .trim_matches('|')
-        .split('|')
-        .map(str::len)
-        .collect();
-    for row in output {
-        let cols: Vec<&str> = row.trim_matches('|').split('|').collect();
-        for (i, col) in cols.iter().enumerate() {
-            assert_eq!(col.len(), widths[i]);
-        }
-    }
-}
-
-#[test]
-fn test_non_table_lines_unchanged() {
-    let input = lines_vec![
-        "# Title",
-        String::new(),
-        "Para text.",
-        String::new(),
-        "| a | b |",
-        "| 1 | 22 |",
-        String::new(),
-        "* bullet",
-        String::new(),
-    ];
-    let output = process_stream(&input);
-    let expected = lines_vec![
-        "# Title",
-        String::new(),
-        "Para text.",
-        String::new(),
-        "| a | b  |",
-        "| 1 | 22 |",
-        String::new(),
-        "* bullet",
-        String::new(),
-    ];
-    assert_eq!(output, expected);
-}
-
-#[test]
-fn test_convert_html_table_basic() {
-    let html_table = lines_vec![
-        "<table>",
-        "<tr><th>A</th><th>B</th></tr>",
-        "<tr><td>1</td><td>2</td></tr>",
-        "</table>",
-    ];
-    let expected = lines_vec!["| A | B |", "| --- | --- |", "| 1 | 2 |"];
-    assert_eq!(convert_html_tables(&html_table), expected);
-}
-
-#[rstest]
-#[case("```")]
-#[case("~~~")]
-#[case("```rust")]
-fn test_convert_html_table_in_text_and_code(#[case] fence: &str) {
-    let lines = lines_vec![
-        "Intro",
-        "<table>",
-        "<tr><th>A</th><th>B</th></tr>",
-        "<tr><td>1</td><td>2</td></tr>",
-        "</table>",
-        fence,
-        "<table><tr><td>x</td></tr></table>",
-        fence,
-        "Outro",
-    ];
-    let expected = lines_vec![
-        "Intro",
-        "| A | B |",
-        "| --- | --- |",
-        "| 1 | 2 |",
-        fence,
-        "<table><tr><td>x</td></tr></table>",
-        fence,
-        "Outro",
-    ];
-    assert_eq!(convert_html_tables(&lines), expected);
-}
-
-#[test]
-fn test_convert_html_table_with_attrs_basic() {
-    let expected = lines_vec!["| A | B |", "| --- | --- |", "| 1 | 2 |"];
-    assert_eq!(convert_html_tables(&html_table_with_attrs()), expected);
-}
-
-#[test]
-fn test_convert_html_table_uppercase() {
-    let expected = lines_vec!["| A | B |", "| --- | --- |", "| 1 | 2 |"];
-    assert_eq!(convert_html_tables(&html_table_uppercase()), expected);
-}
-
-#[test]
-fn test_convert_html_table_with_colspan() {
-    let expected = lines_vec!["| A |", "| --- |", "| 1 | 2 |"];
-    assert_eq!(convert_html_tables(&html_table_with_colspan()), expected);
-}
-
-#[test]
-fn test_convert_html_table_no_header() {
-    let expected = lines_vec!["| A | B |", "| --- | --- |", "| 1 | 2 |"];
-    assert_eq!(convert_html_tables(&html_table_no_header()), expected);
-}
-
-#[test]
-fn test_convert_html_table_empty_row() {
-    let expected = lines_vec!["| 1 | 2 |", "| --- | --- |"];
-    assert_eq!(convert_html_tables(&html_table_empty_row()), expected);
-}
-
-#[test]
-fn test_convert_html_table_whitespace_header() {
-    let expected = lines_vec!["| --- | --- |", "| --- | --- |", "| 1   | 2   |"];
-    assert_eq!(
-        convert_html_tables(&html_table_whitespace_header()),
-        expected
-    );
-}
-
-#[test]
-fn test_convert_html_table_inconsistent_first_row() {
-    let expected = lines_vec!["| A |", "| --- |", "| 1 | 2 |"];
-    assert_eq!(
-        convert_html_tables(&html_table_inconsistent_first_row()),
-        expected
-    );
-}
-
-#[test]
-fn test_convert_html_table_empty() {
-    assert!(convert_html_tables(&html_table_empty()).is_empty());
-}
-
-#[test]
-fn test_convert_html_table_unclosed_returns_original() {
-    let html = html_table_unclosed();
-    assert_eq!(convert_html_tables(&html), html);
-}
-
-#[test]
-fn test_convert_html_table_bold_header() {
-    let input: Vec<String> = include_lines!("data/bold_header_input.txt");
-    let expected: Vec<String> = include_lines!("data/bold_header_expected.txt");
-    assert_eq!(convert_html_tables(&input), expected);
-}
-
-#[test]
-fn test_logical_type_table_output_matches() {
-    let input: Vec<String> = include_lines!("data/logical_type_input.txt");
-    let expected: Vec<String> = include_lines!("data/logical_type_expected.txt");
-    assert_eq!(reflow_table(&input), expected);
-}
-
-#[test]
-/// Verifies that reflowing the option table input produces the expected output.
-///
-/// Loads the input and expected output from external files and asserts that the
-/// `reflow_table` function transforms the input table to match the expected result.
-fn test_option_table_output_matches() {
-    let input: Vec<String> = include_lines!("data/option_table_input.txt");
-    let expected: Vec<String> = include_lines!("data/option_table_expected.txt");
-    assert_eq!(reflow_table(&input), expected);
-}
-
-#[test]
-fn test_month_seconds_table_output_matches() {
-    let input: Vec<String> = include_lines!("data/month_seconds_input.txt");
-    let expected: Vec<String> = include_lines!("data/month_seconds_expected.txt");
-    assert_eq!(reflow_table(&input), expected);
-}
-
-#[test]
-fn test_offset_table_output_matches() {
-    let input: Vec<String> = include_lines!("data/offset_table_input.txt");
-    let expected: Vec<String> = include_lines!("data/offset_table_expected.txt");
-    assert_eq!(reflow_table(&input), expected);
-}
-
-#[test]
-/// Tests that `process_stream` correctly processes a complex Markdown table representing logical
-/// types by comparing its output to expected results loaded from a file.
-fn test_process_stream_logical_type_table() {
-    let input: Vec<String> = include_lines!("data/logical_type_input.txt");
-    let expected: Vec<String> = include_lines!("data/logical_type_expected.txt");
-    assert_eq!(process_stream(&input), expected);
-}
-
-#[test]
-/// Tests that `process_stream` correctly processes a Markdown table with options, producing the
-/// expected output.
-///
-/// Loads input and expected output from test data files, runs `process_stream` on the input, and
-/// asserts equality.
-fn test_process_stream_option_table() {
-    let input: Vec<String> = include_lines!("data/option_table_input.txt");
-    let expected: Vec<String> = include_lines!("data/option_table_expected.txt");
-    assert_eq!(process_stream(&input), expected);
-}
-
-#[test]
-/// Tests that long paragraphs are wrapped at 80 columns by `process_stream`.
-///
-/// Ensures that a single long paragraph is split into multiple lines, each not exceeding 80
-/// characters.
-fn test_wrap_paragraph() {
-    let input = lines_vec![
-        "This is a very long paragraph that should be wrapped at eighty columns so it needs to \
-         contain enough words to exceed that limit.",
-    ];
-    let output = process_stream(&input);
-    assert!(output.len() > 1);
-    assert!(output.iter().all(|l| l.len() <= 80));
-}
-
-#[test]
-fn test_wrap_list_item() {
-    let input = lines_vec![
-        r"- This bullet item is exceptionally long and must be wrapped to keep prefix formatting intact.",
-    ];
-    let output = process_stream(&input);
-    common::assert_wrapped_list_item(&output, "- ", 2);
-}
-
-#[rstest]
-#[case("- ", 3)]
-#[case("1. ", 3)]
-#[case("10. ", 3)]
-#[case("100. ", 3)]
-fn test_wrap_list_items_with_inline_code(#[case] prefix: &str, #[case] expected: usize) {
-    let input = lines_vec![format!(
-        "{prefix}`script`: A multi-line script declared with the YAML `|` block style. The entire \
-         block is passed to an interpreter. If the first line begins with `#!`, Netsuke executes \
-         the script verbatim, respecting the shebang."
-    )];
-    let output = process_stream(&input);
-    common::assert_wrapped_list_item(&output, prefix, expected);
-}
-
-#[test]
-fn test_wrap_preserves_inline_code_spans() {
-    let input = lines_vec![
-        "- `script`: A multi-line script declared with the YAML `|` block style. The entire block \
-         is passed to an interpreter. If the first line begins with `#!`, Netsuke executes the \
-         script verbatim, respecting the shebang.",
-    ];
-    let output = process_stream(&input);
-    common::assert_wrapped_list_item(&output, "- ", 3);
-}
-
-#[test]
-fn test_wrap_multi_backtick_code() {
-    let input = lines_vec![
-        "- ``cmd`` executes ```echo``` output with ``json`` format and prints results to the \
-         console",
-    ];
-    let output = process_stream(&input);
-    common::assert_wrapped_list_item(&output, "- ", 2);
-}
-
-#[test]
-fn test_wrap_multiple_inline_code_spans() {
-    let input = lines_vec![
-        "- Use `foo` and `bar` inside ``baz`` for testing with additional commentary to exceed \
-         wrapping width",
-    ];
-    let output = process_stream(&input);
-    common::assert_wrapped_list_item(&output, "- ", 2);
-}
-#[test]
-fn test_wrap_long_inline_code_item() {
-    let input = lines_vec![concat!(
-        "- `async def on_unhandled(self, ws: WebSocketLike, message: Union[str, bytes])`:",
-        " A fallback handler for messages that are not dispatched by the more specific",
-        " message handlers. This can be used for raw text/binary data or messages that",
-        " don't conform to the expected structured format."
-    )];
-    let output = process_stream(&input);
-    common::assert_wrapped_list_item(&output, "- ", 4);
-    assert!(
-        output
-            .first()
-            .expect("output should not be empty")
-            .ends_with("`:")
-    );
-}
-
-#[test]
-fn test_wrap_future_attribute_punctuation() {
-    let input = vec![
-        concat!(
-            "- Test function (`#[awt]`) or a specific `#[future]` argument ",
-            "(`#[future(awt)]`), tells `rstest` to automatically insert `.await` ",
-            "calls for those futures."
-        )
-        .to_string(),
-    ];
-    let output = process_stream(&input);
-    assert_eq!(
-        output,
-        vec![
-            "- Test function (`#[awt]`) or a specific `#[future]` argument".to_string(),
-            "  (`#[future(awt)]`), tells `rstest` to automatically insert `.await` calls for"
-                .to_string(),
-            "  those futures.".to_string(),
-        ]
-    );
-}
-
-#[test]
-fn test_wrap_footnote_multiline() {
-    let input = lines_vec![concat!(
-        "[^note]: This footnote is sufficiently long to require wrapping ",
-        "across multiple lines so we can verify indentation."
-    )];
-    let output = process_stream(&input);
-    common::assert_wrapped_list_item(&output, "[^note]: ", 2);
-}
-
-#[test]
-fn test_wrap_footnote_with_inline_code() {
-    let input = lines_vec![concat!(
-        "  [^code_note]: A footnote containing inline `code` that should wrap ",
-        "across multiple lines without breaking the span."
-    )];
-    let output = process_stream(&input);
-    common::assert_wrapped_list_item(&output, "  [^code_note]: ", 2);
-}
-
-/// Tests that footnotes with angle-bracketed URLs are wrapped correctly.
-///
-/// Verifies that when a footnote line contains a URL enclosed in angle brackets,
-/// the URL is moved to a new indented line beneath the footnote text.
-#[test]
-fn test_wrap_angle_bracket_url() {
-    let input = lines_vec![concat!(
-        "[^5]: Given When Then - Martin Fowler, accessed on 14 July 2025, ",
-        "<https://martinfowler.com/bliki/GivenWhenThen.html>"
-    )];
-    let expected = lines_vec![
-        "[^5]: Given When Then - Martin Fowler, accessed on 14 July 2025,",
-        "      <https://martinfowler.com/bliki/GivenWhenThen.html>",
-    ];
-    let output = process_stream(&input);
-    assert_eq!(output, expected);
-}
-
-/// Checks that a sequence of footnotes is not altered by wrapping.
-///
-/// This regression test ensures that the footnote collection remains
-/// unchanged when passed to `process_stream`.
-#[test]
-fn test_wrap_footnote_collection() {
-    let input = lines_vec![
-        "[^1]: <https://falcon.readthedocs.io>",
-        "[^2]: <https://asgi.readthedocs.io>",
-        "[^3]: <https://www.starlette.io>",
-        "[^4]: <https://www.starlette.io/websockets/>",
-        "[^5]: <https://channels.readthedocs.io>",
-        "[^6]: <https://channels.readthedocs.io/en/stable/topics/consumers.html>",
-        "[^7]: <https://fastapi.tiangolo.com/advanced/websockets/>",
-        "[^8]: <https://websockets.readthedocs.io>",
-    ];
-
-    let output = process_stream(&input);
-    assert_eq!(output, input);
-}
-
-#[test]
-/// Verifies that short list items are not wrapped or altered by the stream processing logic.
-///
-/// Ensures that a single-line bullet list item remains unchanged after processing.
-fn test_wrap_short_list_item() {
-    let input = lines_vec!["- short item"];
-    let output = process_stream(&input);
-    assert_eq!(output, input);
-}
-
-#[test]
-fn test_wrap_blockquote() {
-    let input = lines_vec![
-        "> **Deprecated**: A :class:`WebSocketRouter` and its `add_route` API should be used to \
-         instantiate resources.",
-    ];
-    let output = process_stream(&input);
-    assert_eq!(
-        output,
-        lines_vec![
-            "> **Deprecated**: A :class:`WebSocketRouter` and its `add_route` API should be",
-            "> used to instantiate resources.",
-        ]
-    );
-}
-
-#[test]
-fn test_wrap_blockquote_nested() {
-    let input = lines_vec![concat!(
-        "> > This nested quote contains enough text to require wrapping so that we ",
-        "can verify multi-level handling."
-    )];
-    let output = process_stream(&input);
-    common::assert_wrapped_blockquote(&output, "> > ", 2);
-    let joined = output
-        .iter()
-        .map(|l| l.trim_start_matches("> > "))
-        .collect::<Vec<_>>()
-        .join(" ");
-    assert_eq!(joined, input[0].trim_start_matches("> > "));
-}
-
-#[test]
-fn test_wrap_blockquote_with_blank_lines() {
-    let input = lines_vec![
-        concat!(
-            "> The first paragraph in this quote is deliberately long enough to wrap ",
-            "across multiple lines so"
-        ),
-        "> demonstrate the behaviour.",
-        ">",
-        concat!(
-            "> The second paragraph is also extended to trigger wrapping in order to ",
-            "ensure blank lines"
-        ),
-        "> are preserved correctly.",
-    ];
-    let output = process_stream(&input);
-    assert_eq!(output[3], ">");
-    common::assert_wrapped_blockquote(&output[..3], "> ", 3);
-    common::assert_wrapped_blockquote(&output[4..], "> ", 3);
-}
-
-#[test]
-fn test_wrap_blockquote_extra_whitespace() {
-    let input = lines_vec![
-        ">    Extra spacing should not prevent correct wrapping of this quoted text that exceeds \
-         the line width.",
-    ];
-    let output = process_stream(&input);
-    common::assert_wrapped_blockquote(&output, ">    ", 2);
-    let joined = output
-        .iter()
-        .map(|l| l.trim_start_matches(">    "))
-        .collect::<Vec<_>>()
-        .join(" ");
-    assert_eq!(joined, input[0].trim_start_matches(">    "));
-}
-
-#[test]
-fn test_wrap_blockquote_short() {
-    let input = lines_vec!["> short"];
-    let output = process_stream(&input);
-    assert_eq!(output, input);
-}
-
-#[test]
-/// Tests that lines with hard line breaks (trailing spaces) are preserved after processing.
-///
-/// Ensures that the `process_stream` function does not remove or alter lines ending with Markdown
-/// hard line breaks.
-fn test_preserve_hard_line_breaks() {
-    let input = lines_vec!["Line one with break.  ", "Line two follows."];
-    let output = process_stream(&input);
-    assert_eq!(output.len(), 2);
-    assert_eq!(output[0], "Line one with break.");
-    assert_eq!(output[1], "Line two follows.");
-}
-
-#[test]
-fn test_wrap_hard_linebreak_backslash() {
-    let input: Vec<String> = include_lines!("data/hard_linebreak_input.txt");
-    let expected: Vec<String> = include_lines!("data/hard_linebreak_expected.txt");
-    assert_eq!(process_stream(&input), expected);
-}
-
-#[test]
-fn test_wrap_hard_linebreak_backslash_edge_cases() {
-    let input = lines_vec!(
-        "This line ends with two backslashes: \\\\",
-        "This line ends with a single backslash: \\",
-        " \\ ",
-        "\\",
-        "Text before \\ and after",
-        "   \\",
-        "",
-    );
-    let expected = lines_vec!(
-        "This line ends with two backslashes: \\\\ This line ends with a single backslash:",
-        "\\",
-        "\\",
-        "\\",
-        "Text before \\ and after \\",
-        "",
-    );
-    assert_eq!(process_stream(&input), expected);
-}
-
-#[test]
-/// Tests that `process_stream` preserves complex table formatting without modification.
-///
-/// This regression test ensures that properly formatted complex tables with multiple
-/// columns and detailed content pass through the processing pipeline unchanged,
-/// preventing regressions that might inadvertently alter correct formatting.
-fn test_regression_complex_table() {
-    let input: Vec<String> = include_lines!("data/regression_table_input.txt");
-    let expected: Vec<String> = include_lines!("data/regression_table_expected.txt");
-    assert_eq!(process_stream(&input), expected);
-}
-
-#[test]
-fn test_renumber_basic() {
-    let input = lines_vec!["1. first", "2. second", "7. third"];
-    let expected = lines_vec!["1. first", "2. second", "3. third"];
-    assert_eq!(renumber_lists(&input), expected);
-}
-
-#[test]
-fn test_renumber_with_fence() {
-    let input = lines_vec!["1. item", "```", "code", "```", "9. next"];
-    let expected = lines_vec!["1. item", "```", "code", "```", "2. next"];
-    assert_eq!(renumber_lists(&input), expected);
-}
-
-#[test]
-fn test_cli_renumber_option() {
-    let output = Command::cargo_bin("mdtablefix")
-        .unwrap()
-        .arg("--renumber")
-        .write_stdin("1. a\n4. b\n")
-        .output()
-        .unwrap();
-    assert!(output.status.success());
-    let text = String::from_utf8_lossy(&output.stdout);
-    assert_eq!(text, "1. a\n2. b\n");
-}
-
-#[test]
-fn test_renumber_nested_lists() {
-    let input = lines_vec![
-        "1. first",
-        "    1. sub first",
-        "    3. sub second",
-        "2. second",
-    ];
-
-    let expected = lines_vec![
-        "1. first",
-        "    1. sub first",
-        "    2. sub second",
-        "2. second",
-    ];
-
-    assert_eq!(renumber_lists(&input), expected);
-}
-
-#[test]
-fn test_renumber_tabs_in_indent() {
-    let input = lines_vec!["1. first", "\t1. sub first", "\t5. sub second", "2. second"];
-
-    let expected = lines_vec!["1. first", "\t1. sub first", "\t2. sub second", "2. second"];
-
-    assert_eq!(renumber_lists(&input), expected);
-}
-
-#[test]
-fn test_renumber_mult_paragraph_items() {
-    let input = lines_vec!["1. first", "", "    still first paragraph", "", "2. second"];
-
-    let expected = lines_vec!["1. first", "", "    still first paragraph", "", "2. second"];
-
-    assert_eq!(renumber_lists(&input), expected);
-}
-
-#[test]
-fn test_renumber_table_in_list() {
-    let input = lines_vec!["1. first", "    | A | B |", "    | 1 | 2 |", "5. second"];
-
-    let expected = lines_vec!["1. first", "    | A | B |", "    | 1 | 2 |", "2. second"];
-
-    assert_eq!(renumber_lists(&input), expected);
-}
-
-#[test]
-fn test_renumber_restart_after_paragraph() {
-    let input: Vec<String> = include_lines!("data/renumber_paragraph_restart_input.txt");
-    let expected: Vec<String> = include_lines!("data/renumber_paragraph_restart_expected.txt");
-    assert_eq!(renumber_lists(&input), expected);
-}
-
-#[test]
-fn test_renumber_restart_after_formatting_paragraph() {
-    let input: Vec<String> = include_str!("data/renumber_formatting_paragraph_input.txt")
-        .lines()
-        .map(str::to_string)
-        .collect();
-    let expected: Vec<String> = include_str!("data/renumber_formatting_paragraph_expected.txt")
-        .lines()
-        .map(str::to_string)
-        .collect();
-    assert_eq!(renumber_lists(&input), expected);
-}
-
-#[test]
-fn test_format_breaks_basic() {
-    let input = lines_vec!["foo", "***", "bar"];
-    let expected: Vec<Cow<str>> = vec![
-        input[0].as_str().into(),
-        Cow::Owned("_".repeat(THEMATIC_BREAK_LEN)),
-        input[2].as_str().into(),
-    ];
-    assert_eq!(format_breaks(&input), expected);
-}
-
-#[test]
-fn test_format_breaks_ignores_code() {
-    let input = lines_vec!["```", "---", "```"];
-    let expected: Vec<Cow<str>> = input.iter().map(|s| s.as_str().into()).collect();
-    assert_eq!(format_breaks(&input), expected);
-}
-
-#[test]
-fn test_format_breaks_mixed_chars() {
-    let input = lines_vec!["-*-*-"];
-    let expected: Vec<Cow<str>> = input.iter().map(|s| s.as_str().into()).collect();
-    assert_eq!(format_breaks(&input), expected);
-}
-
-#[test]
-fn test_format_breaks_with_spaces_and_indent() {
-    let input = lines_vec!["  -  -  -  "];
-    let expected: Vec<Cow<str>> = vec![Cow::Owned("_".repeat(THEMATIC_BREAK_LEN))];
-    assert_eq!(format_breaks(&input), expected);
-}
-
-#[test]
-fn test_format_breaks_with_tabs_and_underscores() {
-    let input = lines_vec!["\t_\t_\t_\t"];
-    let expected: Vec<Cow<str>> = vec![Cow::Owned("_".repeat(THEMATIC_BREAK_LEN))];
-    assert_eq!(format_breaks(&input), expected);
-}
-
-#[test]
-fn test_cli_breaks_option() {
-    let output = Command::cargo_bin("mdtablefix")
-        .unwrap()
-        .arg("--breaks")
-        .write_stdin("---\n")
-        .output()
-        .unwrap();
-    assert!(output.status.success());
-    assert_eq!(
-        String::from_utf8_lossy(&output.stdout),
-        format!("{}\n", "_".repeat(THEMATIC_BREAK_LEN))
-    );
-}
-
-#[test]
-fn test_cli_ellipsis_option() {
-    let output = Command::cargo_bin("mdtablefix")
-        .unwrap()
-        .arg("--ellipsis")
-        .write_stdin("foo...\n")
-        .output()
-        .unwrap();
-    assert!(output.status.success());
-    assert_eq!(String::from_utf8_lossy(&output.stdout), "foo…\n");
-}
-
-#[test]
-fn test_cli_ellipsis_code_span() {
-    let output = Command::cargo_bin("mdtablefix")
-        .unwrap()
-        .arg("--ellipsis")
-        .write_stdin("before `dots...` after\n")
-        .output()
-        .unwrap();
-    assert!(output.status.success());
-    assert_eq!(
-        String::from_utf8_lossy(&output.stdout),
-        "before `dots...` after\n"
-    );
-}
-
-#[test]
-fn test_cli_ellipsis_fenced_block() {
-    let output = Command::cargo_bin("mdtablefix")
-        .unwrap()
-        .arg("--ellipsis")
-        .write_stdin("```\nlet x = ...;\n```\n")
-        .output()
-        .unwrap();
-    assert!(output.status.success());
-    assert_eq!(
-        String::from_utf8_lossy(&output.stdout),
-        "```\nlet x = ...;\n```\n"
-    );
-}
-
-#[test]
-fn test_cli_ellipsis_long_sequence() {
-    let output = Command::cargo_bin("mdtablefix")
-        .unwrap()
-        .arg("--ellipsis")
-        .write_stdin("wait....\n")
-        .output()
-        .unwrap();
-    assert!(output.status.success());
-    assert_eq!(String::from_utf8_lossy(&output.stdout), "wait….\n");
-}
